[Feature #157949153] Remove models from __init__.py
